feat: decouple transaction checks
This commit decouples the transaction checks from transaction controller
so that they can be shared between multiple controllers.  This ensures
safety in case a developers accidentally writes code that uses the
delete interface to delete transactions w/o ensuring the transactions
are properly dereferenced.
